T.T

+++ src/main/java/com/moondust/spleef/content/BattleItemDefinition.java
@@ -9,7 +9,7 @@ public record BattleItemDefinition(
         String action,
         int radius,
         int projectileCount,
-        int fuseSeconds,
+        double fuseSeconds,
         int speedAmplifier,
         int durationSeconds,
         int maxStackSize,
